pyproject.toml
[project]
dependencies = [
  "ccxt>=4.4.43",
  "ipython>=8.31.0",
  "quantstats-lumi>=0.3.3",
  "ruff>=0.11.2",
  "tqdm>=4.67.1",
  "websocket-client>=1.8.0",
]
description = "Add your description here"
name = "test"
readme = "README.md"
requires-python = ">=3.11"
version = "0.1.0"
